Lab2410

+++ Contacts/Application/Services/CategoryServices.cs
@@ -1,18 +1,21 @@
 using Contacts.Application.UseCases;
 using Contacts.Application.UseCases.DTOs;
 using Contacts.Application.Repositories;
-using System.Runtime.Serialization;
+using Contacts.Application; // for IUnitOfWork
+using Contacts.Application.Domain; // for Category
 
 namespace Contacts.Application.Services;
 
 public class CategoryServices(ICategoryRepository categoryRepository, IUnitOfWork unitOfWork) : ICategoryUseCases
 {
-    public void AddCategory(AddCategoryDTO categoryDTO)
+    public AddCategoryResult AddCategory(AddCategoryDTO categoryDTO)
     {
         if (categoryRepository.GetCategoryByName(categoryDTO.Name) == null)
         {
-            categoryRepository.Add(categoryDTO);
+            var category = (Category)categoryDTO;
+            categoryRepository.Add(category);
             unitOfWork.Save();
+            return (AddCategoryResult)category;
         }
         else
         {
@@ -25,12 +28,17 @@ public class CategoryServices(ICategoryRepository categoryRepository, IUnitOfWor
         var category = categoryRepository.GetCategory(categoryID);
         if (category != null)
         {
+            // Prevent deletion if any emails reference this category
+            if (categoryRepository.IsCategoryInUse(categoryID))
+            {
+                throw new ServiceException("Nie można usunąć kategorii, ponieważ przypisane są do niej adresy email");
+            }
             categoryRepository.RemoveCategory(category);
             unitOfWork.Save();
         }
         else
         {
-            throw new ServiceException()
+            throw new ServiceException("Kategoria o podanym ID nie istnieje");
         }
     }
 }
